readyup|refactor: Move the pragma reader beside its only consumer
`declinesFinding` reads an `rdy-ignore` pragma off a source line. Its only consumer is now `resolveFindingOutcome` in the runner, it was never exported from the `readyup/check-utils` barrel, and the README says the pragma is honored by the runner, so a reader following that sentence landed in the directory the reader had just left. It moves to `run/` with its test.

`readSourceText` stays in `check-utils`, where `readTrackedSources` still uses it.
